+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 ### Changed
+## [1.16.5-2.2.1] - 2021-04-26
+### Changed
 - Fabric - Move cloth GUI logic into a separate class so it isn't loaded on the server when config file is read
 - Forge - Fix mod not loading when cloth-config is present with Forge 36.1.10+
 - Fabric - Fixed potential incompatibility with other mods